svec: fix fmc_free_sdb_tree moment
When registering several mezzanines, the second device_register
causes a reprogram that should not do anything (the bitstream
should not change). The release of fmc->sdb information
must happen before, so that fmc_scan_sdb_tree refills
the new SDB information properly.

This is a kludge over a kludge: actually, fmc_scan_sdb_tree
should be called once in the carrier, so that we do not fall into
this.