bugfix: bad vme_addr validation
The new vme_addr sysfs attribute had incorrect validation:
it was the old value stored in the card structure who was
validated by mistake. Moreover, the vme_addr is an offset,
so it has to be compared with the configured size, not
the end of the (absolute) vme address range of the board.