kernel: userspace gateware loading via sysfs
This patch allows to load an userspace-provided bitstream to a SVEC card through sysfs attributes.
It does not rely on kernel firmware loader mechanism (so the gateware files are not restricted
to /lib/firmware directory). Userspace communication is ensured through firmware_blob and firmware_cmd
sysfs entries. The former inputs the bitstream in PAGESIZEd chunks, the latter receives gateware update start/finish commands.