kernel: incompatibility: partial revert 335e1019
The patch 355e1019 move two fields in the spec_dev structure, but other
drivers may include this header and use this structure, so it breakes
all the fields's offset.

Revert this modification in order to make the modification more backward
compatible. It will not fix all the situations (sizeof) but some of them
yes.

Signed-off-by: Federico Vaga <federico.vaga@cern.ch>